ENH: Add Interface.from_class.
Adds a new classmethod, `from_class` to `Interface`, for generating an
interface from an existing class. This is useful when you want to create
a class whose interface matches the interface of an existing class you
can't control (e.g., for creating a test double of class defined in a
library).